fractal-generation: Python tool to benchmark the code
Simple Python script to be called as 'bench/timing.py $REPEATS'. It repeats a set of calls to
'mandelbrot' with '--threads 1' up to '--threads 32', parsing the output timing and storing the
result in a CSV file.

Note that the executions are made in a new folder inside '/tmp', as modern Linux'es mount that in
a RAMdisk. So the overhead for writing the final image should be reduced.